requirements.txt
google-generativeai
pdf2image
Pillow
typing-extensions
